# ordermanagement/openapi.yaml
openapi: 3.0.3
info:
  title: Order Management API
  description: RESTful API for managing orders in vertical slice architecture
  version: 1.0.0

servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://api-staging.example.com/v1
    description: Staging server

paths:
  /orders:
    post:
      summary: Create a new order
      description: Creates a new order in DRAFT status
      operationId: createOrder
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              customerId: "CUST1234"
              items:
                - productId: "PRD00001"
                  productName: "Wireless Headphones"
                  quantity: 2
                  unitPriceCents: 7999
                - productId: "PRD00002"
                  productName: "USB Cable"
                  quantity: 1
                  unitPriceCents: 1299
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found or inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List all orders
      description: Retrieve a paginated list of orders with optional filtering
      operationId: listOrders
      tags:
        - Orders
      parameters:
        - name: customerId
          in: query
          description: Filter by customer ID
          required: false
          schema:
            type: string
            pattern: '^CUST[A-Za-z0-9]{4,8}$'
        - name: status
          in: query
          description: Filter by order status
          required: false
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders/{orderId}:
    get:
      summary: Get order details
      description: Retrieve detailed information about a specific order
      operationId: getOrder
      tags:
        - Orders
      parameters:
        - name: orderId
          in: path
          required: true
          description: Unique identifier for the order
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update order status
      description: Update the status of an existing order
      operationId: updateOrderStatus
      tags:
        - Orders
      parameters:
        - name: orderId
          in: path
          required: true
          description: Unique identifier for the order
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
            example:
              status: "CONFIRMED"
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  #  update the description with test values for customerId and productId
  schemas:
    CreateOrderRequest:
      type: object
      required:
        - customerId
        - items
      properties:
        customerId:
          type: string
          pattern: '^CUST[A-Za-z0-9]{4,8}$'
          description: Customer identifier
          example: "CUST1234"
        items:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: '#/components/schemas/OrderItemRequest'
          description: List of items to include in the order
    OrderItemRequest:
      type: object
      required:
        - productId
        - productName
        - quantity
        - unitPriceCents
      properties:
        productId:
          type: string
          pattern: '^PRD[0-9]{5}'
          example: "PRD12345"
          description: Unique identifier for the product
        productName:
          type: string
          maxLength: 100
          description: Name of the product
        quantity:
          type: integer
          minimum: 1
          maximum: 99
          description: Quantity of the product in the order
        unitPriceCents:
          type: integer
          minimum: 1
          maximum: 999999
          description: Unit price of the product in cents (minimum 1 cent, maximum 9999.99 USD)
    UpdateOrderStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the order
        customerId:
          type: string
          pattern: '^CUST[A-Za-z0-9]{4,8}$'
          description: Customer identifier
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        subtotalCents:
          type: integer
          format: int64
          minimum: 0
          description: Subtotal in cents before tax and shipping
        taxCents:
          type: integer
          format: int64
          minimum: 0
          description: Tax amount in cents (8% of subtotal)
        shippingCents:
          type: integer
          format: int64
          minimum: 0
          description: Shipping cost in cents (500 cents or 0 cents)
        totalCents:
          type: integer
          format: int64
          minimum: 100
          description: Total amount in cents (subtotal + tax + shipping)
        createdAt:
          type: string
          format: date-time
          description: Order creation timestamp
        confirmedAt:
          type: string
          format: date-time
          nullable: true
          description: Order confirmation timestamp
        shippedAt:
          type: string
          format: date-time
          nullable: true
          description: Order shipment timestamp
        deliveredAt:
          type: string
          format: date-time
          nullable: true
          description: Order delivery timestamp
        cancelledAt:
          type: string
          format: date-time
          nullable: true
          description: Order cancellation timestamp

    OrderItemResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the order item
        productId:
          type: string
          pattern: '^PRD[0-9]{5}'

    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/OrderResponse'
        pagination:
          $ref: '#/components/schemas/PaginationResponse'

    PaginationResponse:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        limit:
          type: integer
          minimum: 1
          description: Items per page
        totalItems:
          type: integer
          minimum: 0
          description: Total number of items
        totalPages:
          type: integer
          minimum: 0
          description: Total number of pages
        hasNext:
          type: boolean
          description: Whether there are more pages
        hasPrevious:
          type: boolean
          description: Whether there are previous pages

    OrderStatus:
      type: string
      enum:
        - DRAFT
        - CONFIRMED
        - SHIPPED
        - DELIVERED
        - CANCELLED
      description: Set of valid statuses an order can have during its lifecycle.

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "INVALID_REQUEST"
        message:
          type: string
          description: Human-readable error message
          example: "Order must have at least 1 item"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp